% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.lifetable}
\alias{autoplot.lifetable}
\title{Autoplot a life table}
\usage{
\method{autoplot}{lifetable}(object, ...)
}
\arguments{
\item{object}{A `lifetable` object, or an object inheriting from it.}

\item{...}{Additional arguments passed to `ggplot2::geom_line()`.}
}
\value{
A `ggplot2` plot object.
}
\description{
Produces a ggplot2 representation of the number of survivors by age.
The method is defined for `lifetable`; S3 dispatch also makes it
available for S4 subclasses such as `actuarialtable`.
}
\examples{
lt <- new("lifetable", x = 0:3, lx = c(100, 90, 50, 10))
ggplot2::autoplot(lt)
}
