% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{de_angelis_di_falco}
\alias{de_angelis_di_falco}
\title{Italian Health Insurance Data}
\format{a list containing elevent items (data.frames), and an mdt data object (HealthyMaleTable2013)}
\source{
PAOLO DE ANGELIS, LUIGI DI FALCO (a cura di). Assicurazioni sulla salute: caratteristiche, modelli attuariali e basi tecniche
}
\usage{
de_angelis_di_falco
}
\description{
A list of data.frames containing transition probabilities by age (row) and year of projections
Transitions are split by males and females, and show probabilities of survival, death 
and transitions from Healty to Disabled
}
\keyword{datasets}
