\name{demoChina}
\alias{demoChina}
\docType{data}
\title{
China Mortality Rates for life table construction
}
\description{
Seven yearly mortality rates for each age
}
\usage{data(demoChina)}
\format{
  A data frame with 106 observations on the following 8 variables.
  \describe{
    \item{\code{age}}{Age sequence}
    \item{\code{CL1}}{CL1 table}
    \item{\code{CL2}}{CL2 table}
    \item{\code{CL3}}{a factor with levels \code{0.000340} \code{0.000346} \code{0.000356} \code{0.000374} \code{0.000396} \code{0.000421} \code{0.000457} \code{0.000489} \code{0.000529} \code{0.000579} \code{0.000602} \code{0.000670} \code{0.000702} \code{0.000725} \code{0.000726} \code{0.000728} \code{0.000730} \code{0.000739} \code{0.000743} \code{0.000751} \code{0.000762} \code{0.000767} \code{0.000773} \code{0.000778} \code{0.000781} \code{0.000783} \code{0.000810} \code{0.000855} \code{0.000872} \code{0.000909} \code{0.000976} \code{0.001057} \code{0.001114} \code{0.001146} \code{0.001249} \code{0.001366} \code{0.001470} \code{0.001496} \code{0.001651} \code{0.001812} \code{0.001993} \code{0.002016} \code{0.002192} \code{0.002409} \code{0.002658} \code{0.002909} \code{0.002933} \code{0.003232} \code{0.003557} \code{0.003926} \code{0.004322} \code{0.004770} \code{0.005263} \code{0.005790} \code{0.006367} \code{0.007005} \code{0.007735} \code{0.008524} \code{0.009386} \code{0.010349} \code{0.011378} \code{0.012508} \code{0.013779} \code{0.015168} \code{0.016671} \code{0.018275} \code{0.020108} \code{0.022111} \code{0.024315} \code{0.026702} \code{0.029295} \code{0.032152} \code{0.035305} \code{0.038746} \code{0.042465} \code{0.046582} \code{0.051077} \code{0.055927} \code{0.061236} \code{0.066959} \code{0.073092} \code{0.079823} \code{0.087192} \code{0.095101} \code{0.103652} \code{0.112977} \code{0.123046} \code{0.133930} \code{0.145630} \code{0.158080} \code{0.171596} \code{0.185707} \code{0.200962} \code{0.217248} \code{0.234458} \code{0.253221} \code{0.272360} \code{0.292646} \code{0.314660} \code{0.336476} \code{0.358031} \code{0.381423} \code{0.405431} \code{0.429876} \code{0.454288} \code{1.000000}}
    \item{\code{CL4}}{a factor with levels \code{0.000389} \code{0.000393} \code{0.000412} \code{0.000421} \code{0.000465} \code{0.000468} \code{0.000534} \code{0.000543} \code{0.000620} \code{0.000635} \code{0.000731} \code{0.000739} \code{0.000816} \code{0.000823} \code{0.000826} \code{0.000833} \code{0.000840} \code{0.000851} \code{0.000867} \code{0.000874} \code{0.000883} \code{0.000900} \code{0.000903} \code{0.000906} \code{0.000925} \code{0.000927} \code{0.000943} \code{0.000944} \code{0.000958} \code{0.001022} \code{0.001100} \code{0.001126} \code{0.001189} \code{0.001293} \code{0.001408} \code{0.001450} \code{0.001539} \code{0.001685} \code{0.001846} \code{0.001941} \code{0.002025} \code{0.002223} \code{0.002442} \code{0.002683} \code{0.002733} \code{0.002948} \code{0.003241} \code{0.003562} \code{0.003917} \code{0.004305} \code{0.004734} \code{0.005205} \code{0.005723} \code{0.006291} \code{0.006918} \code{0.007605} \code{0.008358} \code{0.009189} \code{0.010100} \code{0.011100} \code{0.012197} \code{0.013402} \code{0.014725} \code{0.016175} \code{0.017766} \code{0.019509} \code{0.021420} \code{0.023513} \code{0.025804} \code{0.028311} \code{0.031054} \code{0.034051} \code{0.037327} \code{0.040902} \code{0.044801} \code{0.049050} \code{0.053680} \code{0.058713} \code{0.064185} \code{0.070126} \code{0.076561} \code{0.083532} \code{0.091064} \code{0.099197} \code{0.107956} \code{0.117377} \code{0.127486} \code{0.138313} \code{0.149879} \code{0.162214} \code{0.175314} \code{0.189208} \code{0.203896} \code{0.219370} \code{0.235613} \code{0.252637} \code{0.270351} \code{0.288771} \code{0.307840} \code{0.327508} \code{0.347724} \code{0.368255} \code{0.389389} \code{0.410383} \code{0.431881} \code{1.000000}}
    \item{\code{CL5}}{a factor with levels \code{0.000217} \code{0.000223} \code{0.000224} \code{0.000242} \code{0.000246} \code{0.000272} \code{0.000283} \code{0.000307} \code{0.000338} \code{0.000343} \code{0.000379} \code{0.000408} \code{0.000412} \code{0.000433} \code{0.000450} \code{0.000460} \code{0.000466} \code{0.000467} \code{0.000468} \code{0.000473} \code{0.000479} \code{0.000492} \code{0.000509} \code{0.000517} \code{0.000533} \code{0.000563} \code{0.000599} \code{0.000643} \code{0.000660} \code{0.000695} \code{0.000754} \code{0.000823} \code{0.000870} \code{0.000900} \code{0.000989} \code{0.001087} \code{0.001183} \code{0.001198} \code{0.001321} \code{0.001458} \code{0.001611} \code{0.001673} \code{0.001781} \code{0.001970} \code{0.002177} \code{0.002409} \code{0.002488} \code{0.002666} \code{0.002950} \code{0.003265} \code{0.003612} \code{0.003998} \code{0.004425} \code{0.004896} \code{0.005417} \code{0.005995} \code{0.006633} \code{0.007339} \code{0.008120} \code{0.008982} \code{0.009935} \code{0.010989} \code{0.012152} \code{0.013436} \code{0.014855} \code{0.016419} \code{0.018147} \code{0.020050} \code{0.022149} \code{0.024462} \code{0.027008} \code{0.029810} \code{0.032895} \code{0.036281} \code{0.040002} \code{0.044086} \code{0.048562} \code{0.053465} \code{0.058827} \code{0.064688} \code{0.071084} \code{0.078050} \code{0.085629} \code{0.093862} \code{0.102788} \code{0.112440} \code{0.122859} \code{0.134083} \code{0.146138} \code{0.159046} \code{0.172846} \code{0.187522} \code{0.203116} \code{0.219585} \code{0.236961} \code{0.255185} \code{0.274216} \code{0.294027} \code{0.314562} \code{0.335724} \code{0.357486} \code{0.379574} \code{0.402347} \code{1.000000}}
    \item{\code{CL6}}{a factor with levels \code{0.000305} \code{0.000312} \code{0.000320} \code{0.000336} \code{0.000356} \code{0.000379} \code{0.000412} \code{0.000440} \code{0.000476} \code{0.000521} \code{0.000541} \code{0.000603} \code{0.000632} \code{0.000652} \code{0.000655} \code{0.000657} \code{0.000664} \code{0.000669} \code{0.000677} \code{0.000686} \code{0.000690} \code{0.000696} \code{0.000700} \code{0.000702} \code{0.000706} \code{0.000728} \code{0.000770} \code{0.000784} \code{0.000819} \code{0.000878} \code{0.000952} \code{0.001003} \code{0.001031} \code{0.001124} \code{0.001229} \code{0.001323} \code{0.001348} \code{0.001484} \code{0.001631} \code{0.001794} \code{0.001815} \code{0.001973} \code{0.002168} \code{0.002393} \code{0.002618} \code{0.002639} \code{0.002908} \code{0.003202} \code{0.003533} \code{0.003890} \code{0.004293} \code{0.004736} \code{0.005212} \code{0.005729} \code{0.006305} \code{0.006961} \code{0.007671} \code{0.008448} \code{0.009314} \code{0.010240} \code{0.011256} \code{0.012402} \code{0.013650} \code{0.015006} \code{0.016447} \code{0.018096} \code{0.019900} \code{0.021883} \code{0.024031} \code{0.026366} \code{0.028937} \code{0.031775} \code{0.034872} \code{0.038217} \code{0.041924} \code{0.045970} \code{0.050333} \code{0.055113} \code{0.060261} \code{0.065783} \code{0.071841} \code{0.078473} \code{0.085593} \code{0.093287} \code{0.101678} \code{0.110745} \code{0.120533} \code{0.131069} \code{0.142272} \code{0.154437} \code{0.167134} \code{0.180866} \code{0.195528} \code{0.211007} \code{0.227909} \code{0.245119} \code{0.263395} \code{0.283169} \code{0.302822} \code{0.322293} \code{0.343305} \code{0.364759} \code{0.386919} \code{0.409169} \code{1.000000}}
    \item{\code{CL90.93}}{a factor with levels \code{0.000339} \code{0.000346} \code{0.000356} \code{0.000374} \code{0.000396} \code{0.000421} \code{0.000457} \code{0.000489} \code{0.000529} \code{0.000579} \code{0.000602} \code{0.000670} \code{0.000702} \code{0.000724} \code{0.000727} \code{0.000728} \code{0.000730} \code{0.000738} \code{0.000743} \code{0.000752} \code{0.000762} \code{0.000767} \code{0.000773} \code{0.000778} \code{0.000780} \code{0.000784} \code{0.000809} \code{0.000855} \code{0.000872} \code{0.000910} \code{0.000976} \code{0.001057} \code{0.001114} \code{0.001146} \code{0.001249} \code{0.001366} \code{0.001470} \code{0.001497} \code{0.001650} \code{0.001812} \code{0.001993} \code{0.002016} \code{0.002193} \code{0.002409} \code{0.002658} \code{0.002909} \code{0.002933} \code{0.003231} \code{0.003555} \code{0.003925} \code{0.004322} \code{0.004770} \code{0.005263} \code{0.005790} \code{0.006367} \code{0.007005} \code{0.007735} \code{0.008524} \code{0.009386} \code{0.010349} \code{0.011378} \code{0.012508} \code{0.013779} \code{0.015167} \code{0.016672} \code{0.018275} \code{0.020107} \code{0.022111} \code{0.024315} \code{0.026701} \code{0.029296} \code{0.032152} \code{0.035305} \code{0.038746} \code{0.042465} \code{0.046582} \code{0.051078} \code{0.055926} \code{0.061236} \code{0.066958} \code{0.073092} \code{0.079823} \code{0.087192} \code{0.095102} \code{0.103653} \code{0.112976} \code{0.123047} \code{0.133927} \code{0.145631} \code{0.158079} \code{0.171599} \code{0.185702} \code{0.200967} \code{0.217252} \code{0.234450} \code{0.253233} \code{0.272344} \code{0.292664} \code{0.314651} \code{0.336441} \code{0.358080} \code{0.381455} \code{0.405397} \code{0.429801} \code{0.454556} \code{1.000000}}
  }
}
\details{
See the source link for details.
}
\source{
Society of Actuaries mortality web site.
}
\references{
\url{www.mort.soa.org}
}
\examples{
data(demoChina)
tableChinaCL1<-probs2lifetable(probs=demoChina$CL1,radix=1000,type="qx",name="CHINA CL1")
}
\keyword{datasets}
