\name{head-methods}
\docType{methods}
\alias{head-methods}
\alias{head,ANY-method}
\alias{head,lifetable-method}
\title{Methods for Function \code{head} in Package \pkg{lifecontingencies}}
\description{
 This methods returns the head of the \code{data.frame} with columns x and lx-.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "ANY")}}{
A \code{lifetable} object.
}

\item{\code{signature(x = "lifetable")}}{
A \code{lifetable} object.
}
}}
\keyword{methods}

