% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_demographicFunctions.R
\name{mx2qx}
\alias{mx2qx}
\title{Mortality rates to Death probabilities}
\usage{
mx2qx(mx, ax = 0.5)
}
\arguments{
\item{mx}{mortality rates vector}

\item{ax}{the average number of years lived between ages x and x +1 by individuals who die in that interval}
}
\value{
A vector of death probabilities
}
\description{
Function to convert mortality rates to probabilities of death
}
\details{
Function to convert mortality rates to probabilities of death
}
\examples{

#using some recursion
qx2mx(mx2qx(.2))

}
\seealso{
\code{mxt}, \code{qxt}, \code{qx2mx}
}

