\name{print-methods}
\docType{methods}
\alias{print,lifetable-method}
\title{ Print method for \code{lifetable} objects}
\description{
 This functions print a lifetable
}
\section{Methods}{
\describe{
\item{\code{signature(x = "ANY")}}{A lifetable object.}
\item{\code{signature(x = "lifetable")}}{A lifetable}
}}
\keyword{methods}
