% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/3_demographicFunctions.R
\name{qx2mx}
\alias{qx2mx}
\title{Death Probabilities to Mortality Rates}
\usage{
qx2mx(qx, ax = 0.5)
}
\arguments{
\item{qx:}{death probabilities}

\item{ax:}{the average number of years lived between ages x and x +1 by individuals who die in that interval}
}
\value{
A vector of mortality rates
}
\description{
Death Probabilities to Mortality Rates
}
\details{
Function to convert death probabilities to mortality rates
}
\examples{
data(soa08Act)
soa08qx<-as(soa08Act,"numeric")
soa08mx<-qx2mx(qx=soa08qx)
soa08qx2<-mx2qx(soa08mx)
}
\seealso{
\code{mxt}, \code{qxt}, \code{mx2qx}
}

