% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_mdtFunctions.R
\name{qxt.prime.fromMdt}
\alias{qxt.prime.fromMdt}
\title{Return Associated single decrement from absolute rate of decrement}
\usage{
qxt.prime.fromMdt(object, x, t = 1, decrement)
}
\arguments{
\item{object}{a mdj object}

\item{x}{age}

\item{t}{period (default 1)}

\item{decrement}{type (necessary)}
}
\value{
a single value (AST)
}
\description{
Return Associated single decrement from absolute rate of decrement
}
\examples{
Creating the valdez mdf

 valdezDf<-data.frame(
x=c(50:54),
lx=c(4832555,4821937,4810206,4797185,4782737),
hearth=c(5168, 5363, 5618, 5929, 6277),
accidents=c(1157, 1206, 1443, 1679,2152),
other=c(4293,5162,5960,6840,7631))
valdezMdt<-new("mdt",name="ValdezExample",table=valdezDf) 

qxt.prime.fromMdt(object=valdezMdt,x=53,decrement="other")

}
