\name{tail-methods}
\docType{methods}
\alias{tail-methods}
\alias{tail,ANY-method}
\alias{tail,lifetable-method}
\title{Methods for Function \code{tail} in Package \pkg{lifecontingencies}}
\description{
Methods for function \code{tail} in package \pkg{lifecontingencies}
}
\section{Methods}{
\describe{

\item{\code{signature(x = "ANY")}}{
A \code{lifetable} object.
}

\item{\code{signature(x = "lifetable")}}{
A \code{lifetable} object.
}
}}
\keyword{methods}

