% Generated by roxygen2: do not edit by hand
\name{autoplot,lifecontingencies-method}
\alias{autoplot,lifecontingencies-method}
\title{Autoplot a life table}
\usage{
\S4method{autoplot}{lifetable}(object, ...)
}
\arguments{
  \item{object}{A \code{lifetable} object, or an object inheriting from it.}
  \item{...}{Additional arguments passed to \code{ggplot2::geom_line()}.}
}
\value{A \code{ggplot2} plot object.}
\description{Produces a ggplot2 representation of the number of survivors by age. The method is defined for \code{lifetable} and is therefore also available for subclasses such as \code{actuarialtable} through S4 inheritance.}
