\name{coerce-methods}
\docType{methods}
\alias{coerce-methods}
\alias{coerce,actuarialtable,data.frame-method}
\alias{coerce,lifetable,data.frame-method}
\title{Methods for Function \code{coerce}}
\description{
 Function \code{coerce}  transforms lifetable or actuarialtable objects in data.frame S3 objects.
}
\section{Methods}{
\describe{

\item{\code{signature(from = "actuarialtable", to = "data.frame")}}{
\code{coerce} method applied to actuarialtable object.
}

\item{\code{signature(from = "lifetable", to = "data.frame")}}{
\code{coerce} method applied to actuarialtable object.
}
}}
\keyword{methods}

