\name{getOmega-methods}
\docType{methods}
\alias{getOmega-methods}
\alias{getOmega,lifetable-method}
\title{Method to obtain the \eqn{\omega} age within a life table}
\description{
 \code{getOmega}  returns the last attainable age implied in the given \code{lifetable} object.
}
\section{Methods}{
\describe{
\item{\code{signature(object = "lifetable")}}{
A \code{lifetable} object.
}
}}
\keyword{methods}

