% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/8_mdtFunctions.R
\name{Axn.mdt}
\alias{Axn.mdt}
\title{Multiple decrement life insurance}
\usage{
Axn.mdt(object, x, n, i, decrement)
}
\arguments{
\item{object}{an \code{mdt} or \code{actuarialtable} object}

\item{x}{policyholder's age}

\item{n}{contract duration}

\item{i}{interest rate}

\item{decrement}{decrement category}
}
\value{
The scalar representing APV of the insurance
}
\description{
Function to evaluate multiple decrement insurances
}
\section{Warning}{
 The function is experimental and very basic. Testing is still needed. Use at own risk!
}
\examples{
#creates a temporary mdt
myTable<-data.frame(x=41:43,lx=c(800,776,752),d1=rep(8,3),d2=rep(16,3))
myMdt<-new("mdt",table=myTable,name="ciao")
Axn.mdt(myMdt, x=41,n=2,i=.05,decrement="d2")
}

