\name{plot-methods}
\docType{methods}
\alias{plot-methods}
\alias{plot,ANY-method}
\alias{plot,lifetable-method}
\title{Methods for Function \code{plot}}
\description{
 Methods for function \code{plot}
}
\section{Methods}{
\describe{

\item{\code{signature(x = "ANY")}}{
This method shows the survival function implied in actuarialtable and lifetable objects.
}

\item{\code{signature(x = "lifetable")}}{
A lifetable (or actuarial table) object.
}
}}
\keyword{methods}