\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,actuarialtable-method}
\alias{show,lifetable-method}
\title{Methods for Function \code{show}}
\description{
 Methods for function \code{show}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "actuarialtable")}}{
This method plots the Survival function implied in the number at risk per age
in the life table. Following entries are reported:
}

\item{\code{signature(object = "lifetable")}}{
This method plots the Survival function implied in the number at risk per age
in the life table.
}
}}
\keyword{methods}

