% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soaLt}
\alias{soaLt}
\title{Society of Actuaries life table}
\format{
A \code{data.frame} with 111 obs on the following 2 variables:
\describe{ \item{\code{x}}{a numeric vector} \item{\code{Ix}}{a numeric vector} }
}
\usage{
data(soaLt)
}
\description{
This table has been used by the classical book Actuarial Mathematics and by the Society 
of Actuaries for US professional examinations.
}
\details{
Early ages have been found elsewere since miss in the original data sources; SOA did not 
provide population at risk data for certain spans of age (e.g. 1-5, 6-9, 11-14 and 16-19)
}
\examples{
data(soaLt)
head(soaLt)
}
\references{
Actuarial Mathematics (Second Edition), 1997, by Bowers, N.L., Gerber, H.U., 
Hickman, J.C., Jones, D.A. and Nesbitt, C.J.
}
\keyword{datasets}
